memperbaiki transaksi, pembayaran, dll

+++ src/main/java/com/kelompok3/rental_kendaraan_be/controller/MobilController.java
@@ -1,19 +1,27 @@
 package com.kelompok3.rental_kendaraan_be.controller;
 
-import com.kelompok3.rental_kendaraan_be.model.Mobil;
-import com.kelompok3.rental_kendaraan_be.service.MobilService;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-
 import java.io.IOException;
-import java.nio.file.*;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
 import com.kelompok3.rental_kendaraan_be.model.Kendaraan;
+import com.kelompok3.rental_kendaraan_be.model.Mobil;
+import com.kelompok3.rental_kendaraan_be.service.MobilService;
 
 @RestController
 @RequestMapping("/api/mobil")
@@ -37,9 +45,9 @@ public class MobilController {
             @RequestParam("gambar") MultipartFile file,
             @RequestParam("nama") String nama,
             @RequestParam("nomorPolisi") String nomorPolisi,
-            @RequestParam("tahun") int tahun,
+            @RequestParam("tahun") Integer tahun,
             @RequestParam("status") String status,
-            @RequestParam("harga") int harga,
+            @RequestParam("harga") Double harga,
             @RequestParam("jenisTransmisi") String jenisTransmisi,
             @RequestParam("jenisBahanBakar") String jenisBahanBakar,
             @RequestParam("tipeMobil") String tipeMobil,
@@ -58,7 +66,7 @@ public class MobilController {
             mobil.setNomorPolisi(nomorPolisi);
             mobil.setTahun(tahun);
             mobil.setStatus(Kendaraan.StatusKendaraan.valueOf(status));
-            mobil.setHarga( (double) harga);
+            mobil.setHarga(harga);
             mobil.setJenisTransmisi(jenisTransmisi);
             mobil.setJenisBahanBakar(jenisBahanBakar);
             mobil.setTipeMobil(tipeMobil);
@@ -80,7 +88,7 @@ public class MobilController {
             @RequestParam(value = "nomorPolisi", required = false) String nomorPolisi,
             @RequestParam(value = "tahun", required = false) Integer tahun,
             @RequestParam(value = "status", required = false) String status,
-            @RequestParam(value = "harga", required = false) Integer harga,
+            @RequestParam(value = "harga", required = false) Double harga,
             @RequestParam(value = "jenisTransmisi", required = false) String jenisTransmisi,
             @RequestParam(value = "jenisBahanBakar", required = false) String jenisBahanBakar,
             @RequestParam(value = "tipeMobil", required = false) String tipeMobil,
@@ -112,7 +120,7 @@ public class MobilController {
             if (nomorPolisi != null) existing.setNomorPolisi(nomorPolisi);
             if (tahun != null) existing.setTahun(tahun);
             if (status != null) existing.setStatus(Kendaraan.StatusKendaraan.valueOf(status));
-            if (harga != null) existing.setHarga( (double)harga);
+            if (harga != null) existing.setHarga(harga);
             if (jenisTransmisi != null) existing.setJenisTransmisi(jenisTransmisi);
             if (jenisBahanBakar != null) existing.setJenisBahanBakar(jenisBahanBakar);
             if (tipeMobil != null) existing.setTipeMobil(tipeMobil);
